fix-market-info
On Market info section there are harcoded english words "in" on several places , e.g. "in USD" and "in Bitcoin".
I have removed the "in" words, and print out ctrl.selectedMarket selection in brackets, like this for example (USD). Brackets are picked as option so they would't interfere with existing languages, and be universal.
in USD => (USD)
in Bitcoin => (Bitcoin)