[Nanowallet]: upgrade trezor connect to v9
Problem: Trezor login is not working. The dependencies issue caused nem-voting to load improperly, and the wallet is unable to compile and build.
Solution: Implemented Trezor Connect to v9, which resolved the Trezor login issue. Downgrade some dependencies to ensure it can compile in a Node v10 environment. Downgrade nem-library to 1.0.5 because it's the only version compatible with the nem-voting library.