[docs]: Update layout to match updated deps
Changes to GH Pages forced us to update the gh-pages gem,
which in turn updated the just-the-docs theme from 0.3.3 to 0.10.
This required changing the default layout we had.
Fortunately the new layout worked out of the box, and I just
needed to port over a small change: Adding H1 headers to pages.

We also added a bright PREVIEW banner to every page, but I dropped it.
After all these years, I think we can graduate these docs to FINAL :)